+++ modules/backend/src/main/java/eu/bbmri_eric/quality/agent/user/User.java
@@ -1,4 +1,4 @@
-package eu.bbmri_eric.quality.agent.auth;
+package eu.bbmri_eric.quality.agent.user;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
