+++ modules/backend/src/main/java/eu/bbmri_eric/quality/agent/user/UserDetailService.java
@@ -1,4 +1,4 @@
-package eu.bbmri_eric.quality.agent.auth;
+package eu.bbmri_eric.quality.agent.user;
 
 import java.util.List;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
