+++ modules/backend/src/main/java/eu/bbmri_eric/quality/agent/user/UserRepository.java
@@ -1,4 +1,4 @@
-package eu.bbmri_eric.quality.agent.auth;
+package eu.bbmri_eric.quality.agent.user;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
