+++ modules/backend/src/test/java/eu/bbmri_eric/quality/agent/auth/UserDetailServiceIntegrationTests.java
@@ -3,6 +3,9 @@ package eu.bbmri_eric.quality.agent.auth;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import eu.bbmri_eric.quality.agent.user.User;
+import eu.bbmri_eric.quality.agent.user.UserDetailService;
+import eu.bbmri_eric.quality.agent.user.UserRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
